pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "fujulia", email = "juliafu713@gmail.com"},
]
dependencies = [
    "django>=4.2.19",
    "djangorestframework>=3.15.2",
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
